ref: rework the API to nest related types
