+++ Sources/Swift/Integrations/KSCrash/SentryKSCrash+Integration.swift
@@ -1,0 +1,60 @@
+#if ENABLE_KSCRASH
+internal import _SentryPrivate
+internal import KSCrashInstallations
+import Foundation
+
+// MARK: - Integration
+extension SentryKSCrash {
+    typealias DependencyProvider = SentryKSCrash.InstallerProvider
+
+    /// Crash detectors matching SentryCrash's production monitor set:
+    /// Mach exceptions, signals, C++ exceptions, and NSExceptions.
+    /// KSCrash unconditionally adds its Required monitors (System, AppState,
+    /// UserInfo, Resource) on top of whatever is passed here.
+    static let productionSafeMonitors: MonitorType = [
+        .machException,
+        .signal,
+        .cppException,
+        .nsException
+    ]
+
+    final class Integration<Dependencies: DependencyProvider>: NSObject, SwiftIntegration {
+        private weak var options: Options?
+
+        // MARK: - Initialization
+
+        init?(with options: Options, dependencies: Dependencies) {
+            guard options.enableCrashHandler else {
+                SentrySDKLog.debug("Not going to enable \(Self.name) because enableCrashHandler is disabled.")
+                return nil
+            }
+
+            self.options = options
+            super.init()
+
+            do {
+                try dependencies.kscrashInstaller.install(
+                    installPath: options.cacheDirectoryPath,
+                    monitors: productionSafeMonitors.rawValue,
+                    enableSwapCxaThrow: options.experimental.enableUnhandledCPPExceptionsV2
+                )
+            } catch {
+                SentrySDKLog.error("KSCrash install failed: \(error)")
+                return nil
+            }
+
+            SentrySDKInternal.crashReporterInstalled = true
+            if dependencies.kscrashInstaller.crashedLastLaunch {
+                SentrySDKInternal.fatalDetected = true
+            }
+        }
+
+        // MARK: - SwiftIntegration
+        static var name: String {
+            "SentryKSCrashIntegration"
+        }
+
+        func uninstall() {}
+    }
+}
+#endif
